fix speed not changing from dpad input

+++ src/ros/rover/drive/teleop_drive_joy/src/teleop_drive_joy.cpp
@@ -183,7 +183,7 @@ void TeleopDriveJoy::map_button_callbacks()
     switch_controller(DriveMode::DIFF);
   };
 
-  auto changeSpeed = [this](int speed_change)
+  auto changeSpeed = [this](double speed_change)
   {
     if (!locked_)
     {
